fix(lory.js): call reset() *before* dispatching resize event
This fixes a problem when trying to read the index
after a resize in an 'on.lory.resize' handler, where the
old index is returned instead of 0, as set in reset().

Ideally lory.js would not register the resize handler at all
and instead let the application register a resize handler
(maybe even debounced) that calls lory.reset().