fix(lory.js): returnIndex() must factor in the value of infinite
The internal adjustment of the index with the value of infinite is an implementation detail that the
call site should not have to know about. Therefore the value of infinite must be substracted from
the index in returnIndex(). This is in line with slideTo(index) where the given index is adjusted
internally as well, if infinite is set.

BREAKING CHANGE: Any call sites that worked around this bug by manually adjusting the return value
of returnIndex() must remove this work around with this change.